rename recent_rate
