+++ source/consumers/add_ons/CMakeLists.txt
@@ -3,7 +3,7 @@ set(dir ${CMAKE_CURRENT_SOURCE_DIR})
 set(SOURCES
   ${dir}/FilterBlock.cpp
   ${dir}/periodic_trigger.cpp
-  ${dir}/recent_rate.cpp
+  ${dir}/RecentRate.cpp
   ${dir}/Status.cpp
   ${dir}/ValueFilter.cpp
   ${dir}/ValueLatch.cpp
@@ -12,7 +12,7 @@ set(SOURCES
 set(HEADERS
   ${dir}/FilterBlock.h
   ${dir}/periodic_trigger.h
-  ${dir}/recent_rate.h
+  ${dir}/RecentRate.h
   ${dir}/Status.h
   ${dir}/ValueFilter.h
   ${dir}/ValueLatch.h
