+++ source/consumers/add_ons/RecentRate.h
