+++ tests/consumers/add_ons/CMakeLists.txt
@@ -3,7 +3,7 @@ set(dir ${CMAKE_CURRENT_SOURCE_DIR})
 set(SOURCES
   ${dir}/FilterBlockTest.cpp
   ${dir}/periodic_trigger.cpp
-  ${dir}/recent_rate.cpp
+  ${dir}/RecentRateTest.cpp
   ${dir}/StatusTest.cpp
   ${dir}/ValueFilterTest.cpp
   ${dir}/ValueLatchTest.cpp
